@@ -2,7 +2,7 @@ package grmasa.com.open_light;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 public class SplashActivity extends AppCompatActivity {
 
@@ -12,5 +12,5 @@ public class SplashActivity extends AppCompatActivity {
         startActivity(new Intent(SplashActivity.this, MainActivity.class));
         finish();
     }
-    
+
 }
