@@ -3,10 +3,10 @@ package grmasa.com.open_light.fragments;
 import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -70,6 +70,7 @@ public class Device_Fragment extends Fragment  {
         if (activity == null ) {
             return;
         }
+
         Device_fragment_adapter adapter = new Device_fragment_adapter(bulb_ar, getContext());
         lView = view.findViewById(R.id.device_list);
         lView.setAdapter(adapter);
